Improve field directives (#552)
* Redesign Directive Resolvers

The design of Directive field resolvers put burden on implementors of
directives due to the use of the `types` package. Field resolvers which
use receiver types and functions already have a more ergonomic design,
which can be repeated here for directives. This simplifies and clarifies
the implementation of directives for consumers, while putting the burden
of validating types and arguments on the library.

The implementation also incorrectly handled directive arguments
declared in the SDL, which has been corrected.

* Update Directive README example

Reflect the changed directives API usage in the README

* Add directives executable example

Move directives example into main package

Fix directives example

Rename directives example file

Return example file

* Additional Directive Tests and Examples

* Loosen Directive impl Type, allowing alternate visitors

Allowing directives to not have to implement `ResolverInterceptor` in
the future, so long as they implement a valid directive visitor
function. This opens the path for directives which could execute
validation prior to any resolving of fields, without also having to
implement the resolver interceptor function.

While this opens up the type, it is still validated when parsing the
schema. When other functions like validation are supported, those
will also be accepted. This approach avoids a 'no-op' function needing
to be implemented, and the additional calls needing to be made
unnecessarily when resolving fields in the schema.

The names of types have also been updated, and types that don't need to
be exposed have been removed

* Rework directives again to return some safety

Declaring an explicit interface that all directives must implement,
with function `ImplementsDirective`. This is similar to implementing
custom types, where the implementation returns `true` for the types
it implements. This allows the schema option to replace the map with
a varargs slice of directives, localising the knowledge of the name
of the directive within the implementing type.

* Update ImplementsDirective

Don't allow directive visitors to implement multiple directives; they
must only implement one, and return its name. This also allows the
logic for applying directives to be adjusted to only loop through the
visitors once, rather than once per directive declared on the schema

* Ignore non-FIELD_DEFINITION directives

For intercepting field resolution with directive visitors, ignore any
directive location other than FIELD_DEFINITION for now. This is
essential to continue to support other kinds of directives in the
schema, even though directive visitors only implement handling of
field definition currently.

---------

Co-authored-by: pavelnikolov <pavel@section.io>
Co-authored-by: Pavel Nikolov <pavelnikolov@users.noreply.github.com>